[调整CTA基础组件]
