+++ vnpy/app/cta_strategy_pro/__init__.py
@@ -1,18 +1,13 @@
 from pathlib import Path
 
 from vnpy.trader.app import BaseApp
-from vnpy.trader.constant import Direction,Offset,Status,Color
-from vnpy.trader.object import TickData, BarData, TradeData, OrderData
-from vnpy.trader.utility import BarGenerator, ArrayManager
-from .cta_position import CtaPosition
-from .cta_line_bar import CtaLineBar, CtaMinuteBar, CtaHourBar, CtaDayBar, CtaWeekBar
-from .base import APP_NAME, StopOrder, CtaComponent
-from .cta_policy import CtaPolicy
-from .cta_grid_trade import CtaGrid, CtaGridTrade
+from .base import APP_NAME, StopOrder
 
 from .engine import CtaEngine
 
 from .template import CtaTemplate, CtaSignal, TargetPosTemplate, CtaProTemplate, CtaProFutureTemplate
+from .template_spread import CtaSpreadTemplate
+
 
 class CtaStrategyProApp(BaseApp):
     """"""
