+++ vnpy/trader/utility.py
@@ -246,7 +246,7 @@ def get_icon_path(filepath: str, ico_name: str):
     return str(icon_path)
 
 
-def load_json(filename: str, auto_save : bool = True):
+def load_json(filename: str, auto_save: bool = True):
     """
     Load data from json file in temp path.
     """
